fixup! WIP: makepkg git verification
Finish implementation by checking if a repo is signed.

A repo is considered signed if it has a *trailing* url query of the
form: '\?signed$'

Adds two utility functions in util/source.sh for analyzing fragments and
queries, and cleans up source/git.sh to use the utility function(s) when
available.