fixup! WIP: makepkg git verification
Choose the second (later) method of checking for signatures.

It really is just too ugly the other way, even if it happens at the same
time. Also, grep could find a weird commit message too, so...

Also fix the case of files at the same time -- this was actually setting
the filepath, whereas before we originally set the filename.