makepkg: respect --ignorearch when verifying source or printing packagelist
--verifysource is often used to validate the PKGBUILD before uploading
to the AUR, but currently there is no way to trivially check all sources.
By default we don't check sources we won't use, because it forces
downloading those sources, but in certain cases the user might need to
check them regardless...

--packagelist is usually used to derive the built package names for the
user's machine, but in some cases, e.g. repo building, we might want to
get the packagelist for all supported arches. This is primarily
motivated by the desire to use this in the dbscripts testsuite (which
tests for i686 and x86_64) and elsewhere, for automation purposes when
dealing with multi-arch repository building. While one could use `for a
in ${ARCHES[@]}; do CARCH=$a makepkg --packagelist` it seemed easier to
add this directly in makepkg once I was already changing --verifysource.

This partially reverts d8591dd3418d55c5736022ef003891fc03b953e0 but
ensures the use is controllable.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>