repo-add: use more libmakepkg to handle common compression routines
Currently the list of supported formats for an archive, is maintained in
two places. And repo-add does not actually get updated. :(

In the process, remove some of the logical duplication when calling
bsdtar/compress_as. Also, moves to emulating makepkg by not returning a
hard error if given an invalid archive extension; compress_as will
accept it, throw a warning, and proceed using `cat` as the compression
filter, which pacman consumes just as well.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>