makepkg: preserve the path to the bash interpreter through fakeroot
We explicitly configure the shebang path to bash in order to work across
many systems -- some of which do not have bash in /bin. Rather than
allowing the fakeroot invocation to do PATH lookup anyway, explicitly
invoke the same bash interpreter the initial script was running with.
This ("$BASH") is similar to using e.g. sys.executable in Python.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>