doc: describe when and why the prepare function should be used
We don't want people to run architecture/OS/makepkg.conf specific
processes during prepare() and in fact it's been observed that makepkg
by design doesn't even run prepare_buildenv() for it, so the prohibition
against this is now baked into makepkg.

Reflect this differentiation in the documentation on just what, exactly,
a prepare() function is.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>