makepkg: add support for SAMUFLAGS, comparable to MAKEFLAGS
SAMUFLAGS is recently added to the samurai build tool, an implementation
of the ninja build format (which is a competitor to Makefiles).
Currently, samurai implements support for -v (verbose) and -j
(configuring the number of parallel jobs) and rejects all other
SAMUFLAGS.

Adding SAMUFLAGS support to makepkg will allow users to configure how
many jobs to use when building packages, which is an important
consideration as the ninja build format is becoming fairly popular these
days; meson uses it exclusively, cmake can optionally output it, and it
can often speed up the time a build takes.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>