libmakepkg: relax linting rules for unknown PKGBUILD options
Unknown options may be provided by libmakepkg dropins, in which case a
package may want to ensure they are not used. A common case would be the
upx "optimizer", which is broken software that breaks some common
archiving tricks used in the wild, e.g. AppImage containers. Among
things which would actually end up in a pacman package, there is the
Dropbox binary.