makepkg: pass packagelist to pacman -S/-R as stdin
Some users may wish to whitelist pacman in sudoers.d for a builduser,
e.g. makechrootpkg. Currently this requires whitelisting arguments,
because the list of packages is unpredictable; as a result, this is also
whitelisted: -y --config "$PWD"/malicious-pacman.conf

Make it possible to lock down the sudoers policy more thoroughly, as:

pacman -S -
pacman -Rnu -

This will allow makepkg -s to still work as long as the sudoers policy
assumes any package available in the /etc/pacman.conf repositories is
safe to install, and, separately, allow makepkg -r to work as long as
the sudoers policy assumes any package is allowed to be uninstalled.