makepkg: do not save fflags when creating packages
This breaks reproducible builds due to encoding information specific to
the filesystem which was used to build the package, information which we
are never interested in for packaging purposes anyway.

It also means that attempting to extract a package file as root (or
fakeroot) might result in angry warnings being printed to the console by
bsdtar, followed by a non-zero exit code, unless the user remembers to
use --no-fflags during extraction. This is unpleasant UI, even if pacman
itself won't care about these.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>