build: add script to robustly get git version
It's not so simple to do properly, since we don't want to detect random
git repos to which we extracted a plain tarball. GIT_CEILING_DIRECTORIES
lets us ban that.

While we are at it, if the run_command failed, meson does not
automatically make that an error, you need to check .returncode() for
that. So on errors, we used the empty string because stdout was empty.